+++ app/javascript/packages/normalize-yaml/index.js
@@ -2,7 +2,7 @@ import YAML from 'yaml';
 import prettier from 'prettier';
 import { getUnifiedVisitor } from './visitors/index.js';
 
-/** @typedef {'smartPunctuation'|'sortKeys'|'collapseSentenceSpacing'} Formatter */
+/** @typedef {'smartPunctuation'|'sortKeys'|'collapseSpacing'} Formatter */
 
 /**
  * @typedef NormalizeOptions