+++ app/javascript/packages/normalize-yaml/index.spec.js
@@ -22,8 +22,8 @@ describe('normalize', () => {
     expect(await normalize(original)).to.equal(expected);
   });
 
-  it('collapses multiple spaces between sentences', async () => {
-    const original = '---\nparagraph: Lorem ipsum.  Dolor sit amet.';
+  it('collapses multiple spaces', async () => {
+    const original = '---\nparagraph: Lorem ipsum.  Dolor sit  amet.';
     const expected = '---\nparagraph: Lorem ipsum. Dolor sit amet.\n';
 
     expect(await normalize(original)).to.equal(expected);