+++ app/javascript/packages/normalize-yaml/visitors/collapse-spacing.js
@@ -1,7 +1,7 @@
 export default /** @type {import('yaml').visitor} */ ({
   Scalar(_key, node) {
     if (typeof node.value === 'string') {
-      node.value = node.value.replace(/(\w)\. {2,}/g, '$1. ');
+      node.value = node.value.replace(/ {2,}/g, ' ');
     }
   },
 });